Create BSD License
The BSD License is a permissive open-source license that places few limits on how users can use, alter, and distribute software. It is well-liked in the open-source, academic, and business worlds since it is very flexible, has few restrictions, and works with other licenses. The BSD license appeals to developers because it is straightforward, flexible, and compatible with other licenses. Commercially speaking, it enables companies to include the licensed software in their exclusive goods or services without divulging the source code. With a lengthy history and a vibrant user and developer community, the BSD license encourages cooperation and future project contributions. Overall, it has modest restrictions, flexibility, interoperability with other licenses, and a business-friendly outlook.  However, it is important to consider project-specific requirements and the implications of the license's compatibility with third-party components.